Fixed support for C++11 threads
libstdc++-v3 includes a configure test that looks for the internal
gthread library. This configure test decides in part whether C++11
thread support will be built.

As part of this test it tries to include "gtrd.h", which in turn
includes "gtrd-default.h", which is not found by the test since it
only seems to look for it in the configured build tree of libgcc.

So we simply run configure for libgcc so that it generates the needed
headers before building libstdc++.