Make Python build with the sqlite3 module
In order to do this we move Python to build after SQlite3 and
add the hbb_shlib variant to the include and library search paths
of Python. This allows Python to find the SQlite3 library that we
built for linking with shared library, which is what Python needs
in order to build the SQlite3 module.