Fix yum repos for EOL status
The repo configs need to use the baseurl option instead of mirrorlist.
The url is vault.centos.org not mirror.centos.org for EOL OSs.
The path format for vault repos is different, rewrite it to work.
The 64 bit image has a libselinux repo the the 32 bit image does not, support both setups.