Enable C++ std::thread support in libstdc++.a
The libstdc++ build system has a bug. In order for it to enable C++11 thread support,
it checks for gthreads (part of libgcc) support. This is done by checking whether gthr.h
can be found and compiled. gthr.h in turn includes gthr-default.h, which is autogenerated
at the end of the configure script and placed in include/bits.

Therefore we need to run configure twice. The first time to generate include/bits/gthr-default.h,
which allows the second configure run to detect gthreads support.

Closes #19.